requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==3.0
Flask-Login==0.5.0
gunicorn==20.1.0
psycopg2==2.9.1
PyJWT==2.1.0
waitress==2.0.0
